switch to Long addresses and values (d09p1)

+++ src/main/scala/aoc/intcode/Param.scala
@@ -3,15 +3,15 @@ package aoc.intcode
 sealed trait Param
 
 object Param {
-  case class ImmediateParam(value: Int) extends Param {
+  case class ImmediateParam(value: Long) extends Param {
     override def toString = s"$value"
   }
 
-  case class PositionParam(position: Int) extends Param {
+  case class PositionParam(position: Long) extends Param {
     override def toString = s"[$position]"
   }
 
-  case class RelativeBaseParam(offset: Int) extends Param {
+  case class RelativeBaseParam(offset: Long) extends Param {
     override def toString = s"[RB± $offset]"
   }
 }
