+++ src/main/scala/d02/Part1.scala
@@ -3,6 +3,6 @@ package d02
 import aoc.intcode.Program
 
 object Part1 {
-  def patch(p: Program, noun: Int, verb: Int): Program =
+  def patch(p: Program, noun: Long, verb: Long): Program =
     p.write(1, noun).write(2, verb)
 }
