+++ src/main/scala/d02/Part2.scala
@@ -4,13 +4,13 @@ import aoc.intcode.Program
 import cats.implicits._
 
 object Part2 {
-  def result(p: Program): Int = {
+  def result(p: Program): Long = {
     val resultS =
       for {
-        noun <- (0 to 99).toStream
-        verb <- (0 to 99).toStream
+        noun <- (0L to 99L).toStream
+        verb <- (0L to 99L).toStream
         if Part1.patch(p, noun = noun, verb = verb).run.read(0) === 19690720
-      } yield 100 * noun + verb
+      } yield 100L * noun + verb
     resultS.head
   }
 }
