+++ src/main/scala/d05/Part1.scala
@@ -3,6 +3,6 @@ package d05
 import aoc.intcode.Program
 
 object Part1 {
-  def result(p: Program): List[Int] =
+  def result(p: Program): List[Long] =
     p.runOn(List(1))
 }
